Use 'bufio' & string replace. Accepts whitespace strings now.
This is a complete re-work of the previous "scan" method previously used. This new method now uses standard in instead of a 'scan' into memory.

The only problem I have with std in is that it needs a "termination" value. In this case, the stdin is terminated by "\n"

This is problematic because that new line value is interpreted in golangs return or output of that value. I will be wanting to find a way to remove that cleaner value found on line 45.